Set fov_desired maximum value to 130, fix FOV not changing + slider issues
Increases max_fov to allow for up to a value of 130.

This would help players on standard widescreen monitors be able to see as much as players on ultrawide displays, help reduce nausea, to have parity with Valve's other titles such as Left 4 Dead 2 as well as other Portal modding communities who typically allow a higher max FOV setting.

During testing, it was found that changing the fov_desired command and using the slider did not change the FOV whatsoever, which was a separate issue. This has also been resolved with this.